+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog (nionutils)
 
 0.3.21 (UNRELEASED)
 -------------------
+- Filtered lists no longer access their container's master items when closing.
 - Add rotate methods to FloatPoint and FloatSize.
 - Improve LogTicker. Add support for major and minor ticks.
 - Fix case of extending selection with no anchor.